// packages/pro/src/types/reducers/vis-state-updaters.d.ts
export type TooltipField = {
  name: string;
  format: string | null;
};